Orchestrator - Updated planner, orchestrator config, models - Added parallel tests

+++ pkg/models/models.go
@@ -5,7 +5,7 @@ type Task struct {
 	Arg1          float64 `json:"arg1"`
 	Arg2          float64 `json:"arg2"`
 	Operation     string  `json:"operation"`
-	OperationTime int     `json:"operation_time"`
+	OperationTime int64   `json:"operation_time"`
 }
 
 type TaskResult struct {
